feat(courses): implement CourseList component for course display and pagination
