+++ app/courses/page.tsx
@@ -1,26 +1,23 @@
+import CourseList from "@/components/course-list";
 import ActiveFilters from "@/components/search_components/active-filters";
 import Filter from "@/components/search_components/filter";
 import Group from "@/components/search_components/group";
-import Pages from "@/components/search_components/pagination";
 import Sort from "@/components/search_components/sort";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
 import { getColleges } from "@/features/scraping/server/db/repository/college_repository";
-import {
-  getCoursesPage,
-  getTotalPages,
-} from "@/features/scraping/server/db/repository/course_repository";
 import { getDegrees } from "@/features/scraping/server/db/repository/degree_repository";
 import { getDepartments } from "@/features/scraping/server/db/repository/department_repository";
-import { getTranslations } from "next-intl/server";
+import { Suspense } from "react";
+
+function CourseListSkeleton() {
+  return (
+    <div className="py-20 flex flex-col items-center justify-center space-y-4">
+      <div className="w-8 h-8 border-4 border-primary border-t-transparent rounded-full animate-spin" />
+      <p className="text-muted-foreground font-medium animate-pulse">
+        Loading courses...
+      </p>
+    </div>
+  );
+}
 
 export default async function Courses({
   searchParams,
@@ -39,11 +36,14 @@ export default async function Courses({
     getColleges(),
     getDepartments(),
   ]);
-  let dbFilter: Record<string, any>;
+
+  let dbFilter: Record<string, any> = {};
   if (filter) {
-    dbFilter = JSON.parse(Buffer.from(filter, "base64").toString("utf-8"));
-  } else {
-    dbFilter = {};
+    try {
+      dbFilter = JSON.parse(Buffer.from(filter, "base64").toString("utf-8"));
+    } catch (e) {
+      console.error("Failed to parse filter URL parameter");
+    }
   }
 
   const dbParams = {
@@ -51,46 +51,18 @@ export default async function Courses({
     limit: 24,
     filter: dbFilter,
   };
-  const courses = await getCoursesPage(dbParams);
-  const totalPages = await getTotalPages(dbParams);
 
-  let processed = [...courses];
+  const suspenseKey = JSON.stringify(params);
 
-  if (sortBy === "name") {
-    processed.sort((a, b) => a.name.localeCompare(b.name));
-  } else if (sortBy === "hours") {
-    processed.sort((a, b) => b.hours - a.hours);
-  }
-
-  // C. Group
-  let displayData: Record<string, typeof courses> = {};
-
-  if (groupBy === "none") {
-    displayData = { "All Courses": processed };
-  } else {
-    displayData = processed.reduce(
-      (acc, course) => {
-        // @ts-ignore
-        const groupName = course[groupBy].name || "Other";
-
-        if (!acc[groupName]) acc[groupName] = [];
-        acc[groupName].push(course);
-        return acc;
-      },
-      {} as Record<string, typeof courses>,
-    );
-  }
-
-  const t = await getTranslations("Courses")
   return (
     <div className="flex flex-col flex-1">
-      <div className="flex flex-col justify-between gap-5">
+      <div className="flex flex-col justify-between gap-5 mb-5">
         <div className="flex justify-between">
           <Filter
             colleges={colleges}
             degrees={degrees}
             departments={departments}
-          ></Filter>
+          />
           <div className="flex gap-5">
             <Sort />
             <Group />
@@ -102,50 +74,10 @@ export default async function Courses({
           departments={departments}
         />
       </div>
-      {courses.length > 0 ? (
-        <>
-          <Pages totalPages={totalPages} />
-          <div className="py-5 space-y-8">
-            {Object.entries(displayData).map(([groupName, courses]) => (
-              <div key={groupName} className="space-y-4">
-                {groupBy !== "none" && (
-                  <h2 className="text-2xl font-bold border-b pb-2">
-                    {groupName}
-                  </h2>
-                )}
 
-                <div className="grid gap-5 grid-cols-[repeat(auto-fill,minmax(max(350px,30%),1fr))]">
-                  {courses.map((course) => (
-                    <Card key={course._id}>
-                      <CardHeader>
-                        <CardTitle>{course.name}</CardTitle>
-                        <CardDescription>
-                          {course.department.name}
-                        </CardDescription>
-                      </CardHeader>
-                      <CardContent>
-                        <ul className="text-sm text-muted-foreground">
-                          <li>{course.degree.name}</li>
-                          <li>{course.college.name}</li>
-                        </ul>
-                      </CardContent>
-                      <CardFooter className="flex justify-between">
-                        <Badge variant="outline">
-                          {course.hours} Credit Hours
-                        </Badge>
-                        <Button>View Sections</Button>
-                      </CardFooter>
-                    </Card>
-                  ))}
-                </div>
-              </div>
-            ))}
-          </div>
-          <Pages totalPages={totalPages} />
-        </>
-      ) : (
-        <div className="flex justify-center items-center flex-1 text-5xl font-bold">{t("no_courses_found")}</div>
-      )}
+      <Suspense key={suspenseKey} fallback={<CourseListSkeleton />}>
+        <CourseList dbParams={dbParams} sortBy={sortBy} groupBy={groupBy} />
+      </Suspense>
     </div>
   );
 }
